+++ IsekaiMod/Content/Backgrounds/Musician.cs
@@ -2,7 +2,9 @@
 using Kingmaker.Blueprints.Classes;
 using Kingmaker.Blueprints.Classes.Spells;
 using Kingmaker.Designers.Mechanics.Facts;
+using Kingmaker.EntitySystem.Stats;
 using Kingmaker.Enums;
+using Kingmaker.UnitLogic.FactLogic;
 using TabletopTweaks.Core.Utilities;
 using static IsekaiMod.Main;
 
@@ -16,6 +18,12 @@ namespace IsekaiMod.Content.Backgrounds {
                 bp.SetName(IsekaiContext, "Musician");
                 bp.SetBackgroundDescription(IsekaiContext, "The Musician add Persuasion to the list of her class skills and "
                     + "has a +2 competence bonus to caster level for Sonic spells.");
+                bp.AddComponent<AddClassSkill>(c => {
+                    c.Skill = StatType.SkillPersuasion;
+                });
+                bp.AddComponent<AddBackgroundClassSkill>(c => {
+                    c.Skill = StatType.SkillPersuasion;
+                });
                 bp.AddComponent<IncreaseSpellDescriptorCasterLevel>(c => {
                     c.Descriptor = SpellDescriptor.Sonic;
                     c.BonusCasterLevel = 2;